change the name of searchSongPresenter
